ref(symcache): Undeprecate SymCache::functions
