+++ symbolic-debuginfo/src/macho/mod.rs
@@ -804,7 +804,6 @@ impl<'slf, 'd: 'slf> AsSelf<'slf> for MachArchive<'d> {
 mod tests {
     use super::*;
 
-    #[allow(deprecated)]
     #[test]
     fn test_bcsymbolmap() {
         let object_data =
