+++ symbolic-symcache/src/new/compat.rs
@@ -12,33 +12,6 @@ use super::writer::SymCacheConverter;
 use super::*;
 use crate::{SymCacheError, SymCacheErrorKind};
 
-impl<'data> SymCache<'data> {
-    /// An iterator over the functions in this SymCache.
-    pub fn functions(&self) -> Functions<'data> {
-        Functions {
-            cache: self.clone(),
-            function_idx: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Functions<'data> {
-    cache: SymCache<'data>,
-    function_idx: u32,
-}
-
-impl<'data> Iterator for Functions<'data> {
-    type Item = Function<'data>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.cache.get_function(self.function_idx).map(|file| {
-            self.function_idx += 1;
-            file
-        })
-    }
-}
-
 /// A high level writer that can construct SymCaches.
 ///
 /// When using this writer directly, make sure to call [`finish`](SymCacheWriter::finish)
