+++ symbolic-symcache/src/new/lookup.rs
@@ -58,6 +58,17 @@ impl<'data> SymCache<'data> {
             language: Language::from_u32(raw_function.lang),
         })
     }
+
+    /// An iterator over the functions in this SymCache.
+    ///
+    /// Note that functions are *not* returned ordered by name or entry pc,
+    /// but in insertion order, which is essentially random.
+    pub fn functions(&self) -> Functions<'data> {
+        Functions {
+            cache: self.clone(),
+            function_idx: 0,
+        }
+    }
 }
 
 /// A source File included in the SymCache.
@@ -220,5 +231,23 @@ impl<'data, 'cache> Iterator for SourceLocationIter<'data, 'cache> {
                     source_location,
                 }
             })
+    }
+}
+
+/// Iterator returned by [`SymCache::functions`]; see documentation there.
+#[derive(Debug, Clone)]
+pub struct Functions<'data> {
+    cache: SymCache<'data>,
+    function_idx: u32,
+}
+
+impl<'data> Iterator for Functions<'data> {
+    type Item = Function<'data>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.cache.get_function(self.function_idx).map(|file| {
+            self.function_idx += 1;
+            file
+        })
     }
 }
