+++ symbolic-symcache/tests/test_transformers.rs
@@ -1,4 +1,3 @@
-use std::fmt;
 use std::io::Cursor;
 
 use symbolic_common::{ByteView, SelfCell};
@@ -8,36 +7,6 @@ use symbolic_symcache::transform::{self, Transformer};
 use symbolic_symcache::{SymCache, SymCacheWriter};
 
 type Error = Box<dyn std::error::Error>;
-
-/// Helper to create neat snapshots for symbol tables.
-struct FunctionsDebug<'a>(&'a SymCache<'a>);
-
-#[allow(deprecated)]
-impl fmt::Debug for FunctionsDebug<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut vec: Vec<_> = self
-            .0
-            .functions()
-            .filter_map(|f| match f {
-                Ok(f) => {
-                    if f.address() != u32::MAX as u64 {
-                        Some(f)
-                    } else {
-                        None
-                    }
-                }
-                Err(_) => None,
-            })
-            .collect();
-
-        vec.sort_by_key(|f| f.address());
-        for function in vec {
-            writeln!(f, "{:>16x} {}", &function.address(), &function.name())?;
-        }
-
-        Ok(())
-    }
-}
 
 // FIXME: This is a huge pain, can't this be simpler somehow?
 struct OwnedBcSymbolMap(SelfCell<ByteView<'static>, BcSymbolMap<'static>>);
